move timer and change word wrapping lib
