+++ go.mod
@@ -4,7 +4,7 @@ go 1.17
 
 require (
 	github.com/google/go-cmp v0.5.6 // indirect
-	github.com/isbm/textwrap v0.0.0-20190729202254-22edad10bd84
+	github.com/mitchellh/go-wordwrap v1.0.1
 	github.com/pkg/errors v0.9.1 // indirect
 	gotest.tools v2.2.0+incompatible // indirect
 )
